requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
docopt==0.6.2
idna==3.6
Nuitka==2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
ordered-set==4.1.0
pillow==10.2.0
pipreqs==0.4.13
pydub==0.25.1
pygame==2.5.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
requests==2.31.0
SpeechRecognition==3.10.1
typing_extensions==4.9.0
urllib3==2.2.0
yarg==0.1.9
zstandard==0.22.0
